src/requirements.txt
Flask==3.1.0
line-bot-sdk==3.17.1
schedule==1.2.2
python-dotenv==1.1.0
gunicorn==23.0.0
selenium==4.32.0
beautifulsoup4==4.13.4